增加订单表字段以及积分兑换逻辑

+++ src/main/java/com/yihukurama/sysbase/model/OrderEntity.java
@@ -36,14 +36,14 @@ public class OrderEntity extends BaseEntity
 	 **/
 	@ApiModelProperty(value="产生积分")
 	@Column(name="gen_point")
-	private BigDecimal genPoint;
+	private Integer genPoint;
 
 	/**
 	 *消费积分
 	 **/
 	@ApiModelProperty(value="消费积分")
 	@Column(name="consum_point")
-	private BigDecimal consumPoint;
+	private Integer consumPoint;
 
 
 	/**
